requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF processing
pdfplumber==0.10.3

# AI and OpenAI
openai==1.89.0

# Database (for Railway PostgreSQL)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0

# File handling
aiofiles==23.2.1

# Excel export and data processing
pandas>=2.2.0
openpyxl==3.1.2
xlsxwriter==3.1.9 